+++ src/esoteric.h
@@ -35,7 +35,7 @@
 #include "skin.h"
 #include "config.h"
 #include "ui.h"
-#include "ihardware.h"
+#include "hw-ihardware.h"
 
 #ifdef HAVE_LIBOPK
 #include "opkcache.h"