+++ src/hw-linux.h
@@ -5,7 +5,7 @@
 #include <string>
 #include <vector>
 
-#include "ihardware.h"
+#include "hw-ihardware.h"
 #include "sysclock.h"
 #include "constants.h"
 
@@ -43,8 +43,6 @@ class HwLinux : IHardware {
         void ledOff();
 
         int getBatteryLevel();
-        int getVolumeLevel();
-        int setVolumeLevel(int val);
 
         int getBacklightLevel();
         int setBacklightLevel(int val);